Split in two files, with proper visibility for methods on LinearRegression
