+++ examples/linear_regression/linear_regression.rs
@@ -1,9 +1,6 @@
 #![allow(non_snake_case)]
-use ndarray::{stack, Array, Array1, Array2, ArrayBase, Axis, Data, Ix1, Ix2};
-use ndarray_linalg::{random, Solve};
-use ndarray_rand::RandomExt;
-use ndarray_stats::DeviationExt;
-use rand::distributions::StandardNormal;
+use ndarray::{stack, Array, Array1, ArrayBase, Axis, Data, Ix1, Ix2};
+use ndarray_linalg::Solve;
 
 /// The simple linear regression model is
 ///     y = bX + e  where e ~ N(0, sigma^2 * I)
@@ -19,20 +16,20 @@ use rand::distributions::StandardNormal;
 /// where (X^T X)^{-1} X^T is known as the pseudoinverse or Moore-Penrose inverse.
 ///
 /// Adapted from: https://github.com/xinscrs/numpy-ml
-struct LinearRegression {
+pub struct LinearRegression {
     pub beta: Option<Array1<f64>>,
     fit_intercept: bool,
 }
 
 impl LinearRegression {
-    fn new(fit_intercept: bool) -> LinearRegression {
+    pub fn new(fit_intercept: bool) -> LinearRegression {
         LinearRegression {
             beta: None,
             fit_intercept,
         }
     }
 
-    fn fit<A, B>(&mut self, X: ArrayBase<A, Ix2>, y: ArrayBase<B, Ix1>)
+    pub fn fit<A, B>(&mut self, X: ArrayBase<A, Ix2>, y: ArrayBase<B, Ix1>)
     where
         A: Data<Elem = f64>,
         B: Data<Elem = f64>,
@@ -52,19 +49,9 @@ impl LinearRegression {
         };
     }
 
-    fn solve_normal_equation<A, B>(X: ArrayBase<A, Ix2>, y: ArrayBase<B, Ix1>) -> Array1<f64>
-    where
-        A: Data<Elem = f64>,
-        B: Data<Elem = f64>,
-    {
-        let rhs = X.t().dot(&y);
-        let linear_operator = X.t().dot(&X);
-        linear_operator.solve_into(rhs).unwrap()
-    }
-
-    fn predict<A>(&self, X: &ArrayBase<A, Ix2>) -> Array1<f64>
-    where
-        A: Data<Elem = f64>,
+    pub fn predict<A>(&self, X: &ArrayBase<A, Ix2>) -> Array1<f64>
+        where
+            A: Data<Elem = f64>,
     {
         let (n_samples, _) = X.dim();
 
@@ -78,6 +65,16 @@ impl LinearRegression {
         }
     }
 
+    fn solve_normal_equation<A, B>(X: ArrayBase<A, Ix2>, y: ArrayBase<B, Ix1>) -> Array1<f64>
+    where
+        A: Data<Elem = f64>,
+        B: Data<Elem = f64>,
+    {
+        let rhs = X.t().dot(&y);
+        let linear_operator = X.t().dot(&X);
+        linear_operator.solve_into(rhs).unwrap()
+    }
+
     fn _predict<A>(&self, X: &ArrayBase<A, Ix2>) -> Array1<f64>
     where
         A: Data<Elem = f64>,
@@ -87,40 +84,4 @@ impl LinearRegression {
             Some(beta) => X.dot(beta),
         }
     }
-}
-
-fn get_data(n_samples: usize, n_features: usize) -> (Array2<f64>, Array1<f64>) {
-    let shape = (n_samples, n_features);
-    let noise: Array1<f64> = Array::random(n_samples, StandardNormal);
-
-    let beta: Array1<f64> = random(n_features) * 10.;
-    println!("Beta used to generate target variable: {:.3}", beta);
-
-    let X: Array2<f64> = random(shape);
-    let y: Array1<f64> = X.dot(&beta) + noise;
-    (X, y)
-}
-
-pub fn main() {
-    let n_train_samples = 5000;
-    let n_test_samples = 1000;
-    let n_features = 3;
-
-    let (X, y) = get_data(n_train_samples + n_test_samples, n_features);
-    let (X_train, X_test) = X.view().split_at(Axis(0), n_train_samples);
-    let (y_train, y_test) = y.view().split_at(Axis(0), n_train_samples);
-
-    let mut linear_regressor = LinearRegression::new(false);
-    linear_regressor.fit(X_train, y_train);
-
-    let test_predictions = linear_regressor.predict(&X_test);
-    let mean_squared_error = test_predictions.mean_sq_err(&y_test.to_owned()).unwrap();
-    println!(
-        "Beta estimated from the training data: {:.3}",
-        linear_regressor.beta.unwrap()
-    );
-    println!(
-        "The fitted regressor has a mean squared error of {:.3}",
-        mean_squared_error
-    );
 }
